+++ tests/classes/MockRoute.php
@@ -2,6 +2,8 @@
 
 namespace Tivins\WebappTests\classes;
 
+use Tivins\Webapp\RouteAttribute;
+use Tivins\Webapp\ContentType;
 use Tivins\Webapp\HTTPResponse;
 use Tivins\Webapp\Request;
 use Tivins\Webapp\RouteInterface;
@@ -11,6 +13,11 @@ use Tivins\Webapp\RouteInterface;
  */
 class MockRoute implements RouteInterface
 {
+    #[RouteAttribute(
+        name: 'Test route',
+        description: 'Test route used for testing',
+        contentType: ContentType::JSON
+    )]
     public function trigger(Request $request, array $matches): HTTPResponse
     {
         return new HTTPResponse(code: 200, body: ['matches' => $matches, 'path' => $request->path]);